% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agg}
\alias{agg}
\title{Example Data for Non-BIBD Aggregate-Level Best-Worst Scaling}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 4 columns.}
\usage{
agg
}
\description{
A tibble that shows the format a tibble should take when it is
  submitted to an aggregate scoring function that is NOT ae_mnl(), which
  assumes a balanced incomplete block design (BIBD).
}
\keyword{datasets}
