% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_bayescoring.R
\name{e_bayescoring}
\alias{e_bayescoring}
\title{Empirical Bayes Method to Calculate Individual Best-Worst Scores}
\usage{
e_bayescoring(data, id, item, choice, E = 0.01, alpha = 0.1,
  wide = FALSE)
}
\arguments{
\item{data}{A data.frame of the type described in details.}

\item{id}{A string of the name of the id column.}

\item{item}{A string of the name of the item column.}

\item{choice}{A string of the name of the choice column.}

\item{E}{TODO add description}

\item{alpha}{TODO add description}

\item{wide}{Logical of whether or not one wants the data returned in long
(each row is an item-respondent combination and all best-worst scores are
in the same column) format (FALSE) or in wide format (where each row is a 
respondent, and the best-wost scores for the items are in their own 
columns). See the `indiv` data as an example.}
}
\value{

}
\description{

}
\details{
This function requires data to be in a specified format. Each row must
  represent a respondent-block-label combination. That is, it indicates
  the person, the block (or trial), the item that was judged, and a column
  indicating whether it was chosen as best (+1), worst (-1), or wasn't 
  selected as either (0).
}
\examples{


}
\references{

}
