% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checks.R
\name{get_checks}
\alias{get_checks}
\title{Perform Checks That Data Follows Best Worst Scaling Design}
\usage{
get_checks(data, id, block, item, choice, aggregate = FALSE)
}
\arguments{
\item{data}{A data.frame of the type described in details.}

\item{id}{A string of the name of the id column.}

\item{block}{A string of the name of the block column.}

\item{item}{A string of the name of the item column.}

\item{choice}{A string of the name of the choice column.}
}
\description{
Perform Checks That Data Follows Best Worst Scaling Design
}
\details{
This function requires data to be in a specified format. Each row must
  represent a respondent-block-label combination. That is, it indicates
  the person, the block (or trial), the item that was judged, and a column
  indicating whether it was chosen as best (+1), worst (-1), or wasn't 
  selected as either (0).
}
