% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_choice_p.R
\name{get_choice_p}
\alias{get_choice_p}
\title{Get Analytical Choice Probability, Given a Vector of Utility Coefficients}
\usage{
get_choice_p(b)
}
\arguments{
\item{b}{A numeric vector of utility coefficients}
}
\value{
A vector of analytically-estimated utility coefficients
}
\description{
This takes a vector of utility coefficients and returns the same length
vector of analytical probabilities, using equation 18 from
Lipovetsky & Conklin (2014).
}
\references{
Lipovetsky, S., & Conklin, M. (2014). Best-worst scaling in
analytical closed-form solution. The Journal of Choice Modelling, 10,
60-68. doi: 10.1016/j.jocm.2014.02.001
}
