% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eloresults.R
\name{get_eloresults}
\alias{get_eloresults}
\title{Transform Tidy Data into Winners and Losers Data}
\usage{
get_eloresults(data, block, item, choice)
}
\arguments{
\item{data}{A data.frame of the type described in details.}

\item{block}{A string of the name of the block column.}

\item{item}{A string of the name of the item column.}

\item{choice}{A string of the name of the choice column.}
}
\value{
A tibble with columns of winners and losers to score.
}
\description{
Takes tidy data and prepares it for Elo scoring in get_eloscores()
}
