% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eloscores.R
\name{get_eloscores}
\alias{get_eloscores}
\title{Turn Win-Loss Record Data into Elo Scores}
\usage{
get_eloscores(eloresults, K = 30, iter = 100)
}
\arguments{
\item{eloresults}{What get_eloresults() returns}

\item{K}{The Elo K-factor. The default is 30, per Hollis (2018).}

\item{iter}{Number of different randomizations of the "matchup" order to 
iterate through.}
}
\value{
A tibble with columns for the item and Elo score
}
\description{
Takes what get_eloresults() returns and turns these into Elo scores.
}
