% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prscores.R
\name{get_prscores}
\alias{get_prscores}
\title{Get Page Rank from Adjacency Matrix}
\usage{
get_prscores(M, ...)
}
\arguments{
\item{M}{An adjacency matrix, such as produced by get_M.}

\item{...}{Arguments to igraph::page_rank}
}
\value{
A named numeric vecor indicating page rank score.
}
\description{
Takes adjacency matrix and returns raw page rank scores.
}
