% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_walkscores.R
\name{get_walkscores}
\alias{get_walkscores}
\title{Perform Random Walks on Adjacency Matrix}
\usage{
get_walkscores(M, walks)
}
\arguments{
\item{M}{An adjacency matrix, such as produced by get_M.}

\item{walks}{Integer indicating how many random walks to simulate.}
}
\value{
A named numeric vecor indicating proportion of time a walk led to each node.
}
\description{
Takes adjacency matrix and returns raw random walk scores.
}
