% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv}
\alias{indiv}
\title{Example Data for Individual-Level Best-Worst Scoring}
\format{An object of class \code{data.frame} with 72 rows and 4 columns.}
\usage{
indiv
}
\description{
A data.frame that shows the format a data.frame should take when it is
  submitted to the `indiv_bws` function.
}
\keyword{datasets}
