% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv_bws.R
\name{indiv_bws}
\alias{indiv_bws}
\title{Calculate Individual Best-Worst Scores}
\usage{
indiv_bws(data, id, item, choice, wide = FALSE)
}
\arguments{
\item{wide}{}
}
\value{

}
\description{

}
\details{

}
\examples{

}
\references{

}
