% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdata}
\alias{vdata}
\title{Example Data Used in Vignettes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18200 rows and 4 columns.
}
\usage{
vdata
}
\description{
A tibble that shows the format a tibble should take when it is
  submitted to a individual `scoring` function.
}
\keyword{datasets}
