% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkscoring.R
\name{walkscoring}
\alias{walkscoring}
\title{Walkscoring Method to Calculate Individual Best-Worst Scores}
\usage{
walkscoring(data, id, block, item, choice, walks = 10000)
}
\arguments{
\item{data}{A data.frame of the type described in details.}

\item{id}{A string of the name of the id column.}

\item{block}{A string of the name of the block column.}

\item{item}{A string of the name of the item column.}

\item{choice}{A string of the name of the choice column.}

\item{walks}{Integer indicating how many random walks to simulate.}
}
\value{
A data.frame containing the id and columns for each item, containing that
  individual's walkscore for that item.
}
\description{
Caclulate best-worst scores for each respondent-item combination. This uses
  the walkscoring method describe in CITE.
}
\details{
This function requires data to be in a specified format. Each row must
  represent a respondent-block-label combination. That is, it indicates
  the person, the block (or trial), the item that was judged, and a column
  indicating whether it was chosen as best (+1), worst (-1), or wasn't 
  selected as either (0).
}
\examples{
data(indiv)
head(indiv)
# use more than 10 walks; only using 10 here for speed
walkscoring(indiv, "id", "block", "label", "value", 10)

}
\references{
INSERT MY PAPER HERE LATER.
}
