% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiv}
\alias{indiv}
\title{Example Data for Individual-Level Best-Worst Scaling}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 520 rows and 4 columns.}
\usage{
indiv
}
\description{
A tibble that shows the format a tibble should take when it is
  submitted to a individual `scoring` function.
}
\keyword{datasets}
