% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bibd.R
\name{make_bibd}
\alias{make_bibd}
\title{Make Balanced Incomplete Block Designs from bibds Designs}
\usage{
make_bibd(design)
}
\arguments{
\item{design}{Integer from 1 to 32. Corresponds to the characteristics from
the bibds data.frame object.}
}
\value{
A tibble. The first column indicates the block, and the rest of the
  columns indicate which item is in each block.
}
\description{
This function generates a balanced incomplete block design. It takes one
  argument, the design number from the bibds data.frame object. See bibds
}
